engine: extract esm/viz_layer.js
